Move WidgetExt to kas::adapter module
